Extract snapshot infrastructure into ketchup-snapshots gem
Move the generic browser control, diff/comparison, and file I/O
into gems/ketchup-snapshots. The main app's snapshots.rb now
reopens Capture to add the ketchup-specific capture script and
server setup.
